refactor: moved not generic components close to where they are being used
