+++ react-client/src/routes/my-pets/:petId/PetProfileIndex.tsx
@@ -4,7 +4,7 @@ import { Header } from "~/components/design-system/header/Header";
 import { Layout } from "~/components/design-system/layout/Layout";
 import { PetAlertMessage } from "~/components/Pet/PetAlertMessage";
 import { PetCard } from "~/components/Pet/PetCard";
-import { PetCardInfo } from "~/components/Pet/PetCardInfo";
+import { PetCardInfo } from "./components/PetCardInfo";
 import { usePetProfileContext } from "./usePetProfileLayoutViewModel";
 
 export const PetProfileIndex = () => {
